+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.6] - 2025-07-30
+
+### Added
+
+- `Int` now supports:
+  - `Shl<u32, Output = Self>` and `ShlAssign<u32>`
+  - `Shr<u32, Output = Self>` and `ShrAssign<u32>`
+  - `abs` method
+  - `trailing_zeros` method
+- Added `ixy::ops::{distance, line}` functions
+- Added `Pos::normalized_approx` method for approximating a direction vector
+
 ## [0.5.5] - 2025-07-30
 
 ### Added